+++ huffman.h
@@ -15,10 +15,10 @@ typedef struct HTNode
 
 bool operator>(HTNode a, HTNode b);
 
-void init(HuffmanTree &HT, std::map<int, int> &map, char **&HC, int tree_n);
-void init_big(HuffmanTree &HT, std::map<int, int> &map, std::unordered_map<int, char *> &HC, int tree_n);
-void init_for_de(HuffmanTree &HT, std::map<int, int> &map, int tree_n);
+void init(HuffmanTree &HT, std::map<long long int, int> &map, char **&HC, int tree_n);
+void init_big(HuffmanTree &HT, std::map<long long int, int> &map, std::unordered_map<long long int, char *> &HC, int tree_n);
+void init_for_de(HuffmanTree &HT, std::map<long long int, int> &map, int tree_n);
 void select(HuffmanTree &HT, int n);
-void output_huffmantree(HuffmanTree &HT, std::map<int, int> &map, int tree_n);
+void output_huffmantree(HuffmanTree &HT, std::map<long long int, int> &map, int tree_n);
 
 int judge(int a);
